refactor: unify account init and term controllers/subroutines
On-behalf-of: @SAP <simon@simontesar.com>
